# SoftwareQuizAPI/SoftwareQuizAPI/Controllers/QuizController.cs
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using SoftwareQuizAPI.Data;
using SoftwareQuizAPI.Models;

namespace SoftwareQuizAPI.Controllers
{
    [Route("api/QuizController")]
    [ApiController]
    public class QuizController : ControllerBase
    {


        private readonly AppDbContext _context;

        public QuizController(AppDbContext context)
        {
            _context = context;
        }


        [HttpGet]
        public async Task<ActionResult<IEnumerable<Question>>> GetQuestions()
        {
            return await _context.Questions.ToListAsync();
        }

        [HttpGet("{id}")]
        public async Task<ActionResult<Question>> GetQuestion(int id)
        {
            var question = await _context.Questions.FindAsync(id);

            if (question == null)
            {
                return NotFound();
            }

            return question;
        }


        [HttpPost]
        public async Task<ActionResult<Question>> PostQuestion(Question question)
        {
            _context.Questions.Add(question);
            await _context.SaveChangesAsync();

            return CreatedAtAction(nameof(GetQuestion), new { id = question.Id }, question);
        }


        [HttpGet("category/{category}")]
        public async Task<ActionResult<IEnumerable<Question>>> GetQuestionsByCategory(string category)
        {
            var questions = await _context.Questions
                .Where(q => q.Category == category)
                .ToListAsync();

            if (questions == null || !questions.Any())
            {
                return NotFound();
            }

            return questions;
        }

        [HttpPut("{id}")]
        public async Task<IActionResult> PutQuestion(int id, Question question)
        {
            if (id != question.Id)
            {
                return BadRequest();
            }

            _context.Entry(question).State = EntityState.Modified;

            try
            {
                await _context.SaveChangesAsync();
            }
            catch (DbUpdateConcurrencyException)
            {
                if (!QuestionExists(id))
                {
                    return NotFound();
                }
                else
                {
                    throw;
                }
            }

            return NoContent();
        }

        [HttpDelete("{id}")]
        public async Task<IActionResult> DeleteQuestion(int id)
        {
            var question = await _context.Questions.FindAsync(id);
            if (question == null)
            {
                return NotFound();
            }

            _context.Questions.Remove(question);
            await _context.SaveChangesAsync();

            return NoContent();
        }


        private bool QuestionExists(int id)
        {
            return _context.Questions.Any(e => e.Id == id);
        }


    }
}